+++ e2e/src/browser/mock/websocket.rs
@@ -85,12 +85,12 @@ impl<'a> WebSocket<'a> {
         self.0
             .execute(Statement::new(
                 // language=JavaScript
-                r#"
+                "
                 async () => {
                     window.wsMock.isClosed = false;
                     window.wsMock.closeCode = 0;
                 }
-                "#,
+                ",
                 [],
             ))
             .await